Protecting user data

+++ src/Grewbit/GrewbitWeb/Controllers/PlotController.cs
@@ -6,6 +6,7 @@ using System.Web;
 using System.Web.Mvc;
 using GrewbitShared.Data;
 using GrewbitShared.Models;
+using Microsoft.AspNet.Identity;
 
 namespace GrewbitWeb.Controllers
 {
@@ -20,14 +21,20 @@ namespace GrewbitWeb.Controllers
 
         public ActionResult Index()
         {
-            IList<Plot> plots = _plotRepository.GetList();
+            var userId = User.Identity.GetUserId();
+
+            IList<Plot> plots = _plotRepository.GetList(userId);
 
             return View(plots);
         }
 
         public ActionResult Add()
         {
-            return View();
+            var plot = new Plot();
+
+            plot.UserId = User.Identity.GetUserId();
+
+            return View(plot);
         }
 
         [HttpPost]
@@ -36,6 +43,7 @@ namespace GrewbitWeb.Controllers
         {
             if (ModelState.IsValid)
             {
+                plot.UserId = User.Identity.GetUserId();
                 _plotRepository.Add(plot);
 
                 TempData["Message"] = "Plot was successfully added!";
@@ -53,7 +61,9 @@ namespace GrewbitWeb.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
 
-            Plot plot = _plotRepository.Get(id.Value);
+            var userId = User.Identity.GetUserId();
+
+            Plot plot = _plotRepository.Get(id.Value, userId);
 
             if (plot == null)
             {
@@ -70,7 +80,9 @@ namespace GrewbitWeb.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
 
-            Plot plot = _plotRepository.Get(id.Value);
+            var userId = User.Identity.GetUserId();
+
+            Plot plot = _plotRepository.Get(id.Value, userId);
 
             if (plot == null)
             {
@@ -86,6 +98,15 @@ namespace GrewbitWeb.Controllers
         {
             if (ModelState.IsValid)
             {
+                var userId = User.Identity.GetUserId();
+
+                if (!_plotRepository.PlotOwnedByUserId(plot.Id, userId))
+                {
+                    return HttpNotFound();
+                }
+
+                plot.UserId = userId;
+
                 _plotRepository.Update(plot);
 
                 TempData["Message"] = "Plot was successfully updated!";
@@ -100,6 +121,13 @@ namespace GrewbitWeb.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Delete(int? id)
         {
+            var userId = User.Identity.GetUserId();
+
+            if (!_plotRepository.PlotOwnedByUserId(id.Value, userId))
+            {
+                return HttpNotFound();
+            }
+
             _plotRepository.Delete(id.Value);
 
             TempData["Message"] = "Plot was successfully deleted!";
